feat: 調整路由, 新增模板、模型取得資料
- 調整 authentication 路由，用於 login 和 registration。
- 重構主應用程式路由。
- 重新命名 optimizer 路由，並將其功能合併到主應用程式中。
- 更新 route tree，包含新的 auth 路由並移除過時的路由。
- 新增 models 和 templates 的新 types，以提升 type safety。
- 增強 UI components，提升 authentication 和 template 管理的用戶體驗。
- 新增 models 和 templates 取得列表 API。
