+++ frontend/src/hooks/useAppSession.ts
@@ -1,0 +1,8 @@
+import type { AuthResponse } from '@/types/auth'
+import { useSession } from '@tanstack/react-start/server'
+
+export function useAppSession() {
+  return useSession<AuthResponse>({
+    password: 'ChangeThisBeforeShippingToProdOrYouWillBeFired',
+  })
+}
