+++ frontend/src/routes/__root.tsx
@@ -1,6 +1,7 @@
 // app/routes/__root.tsx
 import type { QueryClient } from '@tanstack/react-query'
 import type { ReactNode } from 'react'
+import type { AuthResponse } from '@/types/auth'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import {
   createRootRouteWithContext,
@@ -12,9 +13,12 @@ import { TanStackRouterDevtools } from '@tanstack/react-router-devtools'
 import { Toaster } from '@/components/ui/sonner'
 import appCss from '@/styles/app.css?url'
 
-export const Route = createRootRouteWithContext<{
+interface RouterContext {
   queryClient: QueryClient
-}>()({
+  auth?: AuthResponse
+}
+
+export const Route = createRootRouteWithContext<RouterContext>()({
   head: () => ({
     meta: [
       {
