+++ frontend/src/types/model.ts
@@ -1,0 +1,8 @@
+export interface ModelEntry {
+  name: string
+  displayName: string
+  description: string
+  version: string
+}
+
+export type ModelEntries = Array<ModelEntry>
