+++ frontend/src/types/optimize.ts
@@ -1,0 +1,7 @@
+export interface OptimizePayload {
+  original_prompt: string
+  template_id: number
+  model: string
+  temperature: number
+  max_output_tokens?: number
+}
