+++ frontend/src/utils/api.ts
@@ -1,0 +1,5 @@
+import { ofetch } from 'ofetch'
+
+export const api = ofetch.create({
+  baseURL: import.meta.env.VITE_API_URL,
+})
